CLI - refactoring
