+++ fittrackee/cli/app.py
@@ -1,0 +1,3 @@
+from fittrackee import create_app
+
+app = create_app()
