+++ pyproject.toml
@@ -63,9 +63,8 @@ Sphinx = "^4.5"
 
 [tool.poetry.scripts]
 fittrackee = 'fittrackee.__main__:main'
-fittrackee_set_admin = 'fittrackee.users.users:set_admin'
-fittrackee_upgrade_db = 'fittrackee.__main__:upgrade_db'
 fittrackee_worker = 'flask_dramatiq:worker'
+ftcli = 'fittrackee.cli:cli'
 
 [tool.black]
 line-length = 79
